+++ tests/acceptance/fixtures/responses/cors.js
@@ -1,0 +1,1 @@
+export const CORSHeaders = { 'access-control-allow-credentials': true, 'access-control-allow-origin': 'http://localhost:9999' };
