+++ tests/acceptance/fixtures/responses/universal/search.js
@@ -1,43 +1,5 @@
 import { RequestMock } from 'testcafe';
-
-function generateAutoCompleteResponse (prompt) {
-  const mockedResponse = {
-    meta: {
-      uuid: '01802d71-9901-1b83-9d50-ff143088f1ab',
-      errors: []
-    },
-    response: {
-      input: {
-        value: prompt,
-        queryIntents: []
-      },
-      results: []
-    }
-  };
-
-  if (prompt === '') {
-    mockedResponse.response.results = [
-      {
-        value: 'a Rose by any other name',
-        matchedSubstrings: [],
-        queryIntents: [],
-        verticalKeys: []
-      },
-      {
-        value: 'amani farooque phone number',
-        matchedSubstrings: [],
-        queryIntents: [],
-        verticalKeys: []
-      }
-    ];
-  } else if (prompt.startsWith('a')) {
-    mockedResponse.response.results = [
-      { value: 'amani farooque phone number', matchedSubstrings: [], queryIntents: [], verticalKeys: [] }
-    ];
-  }
-
-  return mockedResponse;
-}
+import { CORSHeaders } from '../cors';
 
 const UniversalSearchResponse = {
   meta: {
@@ -515,7 +477,6 @@ const UniversalSearchResponse = {
     }
   }
 };
-const CORSHeaders = { 'access-control-allow-credentials': true, 'access-control-allow-origin': 'http://localhost:9999' };
 
 export const MockedUniversalSearchRequest = RequestMock()
   .onRequestTo(async request => {
@@ -523,15 +484,3 @@ export const MockedUniversalSearchRequest = RequestMock()
     return urlRegex.test(request.url) && request.method === 'get';
   })
   .respond(UniversalSearchResponse, 200, CORSHeaders);
-
-export const MockedUniversalAutoCompleteRequest = RequestMock()
-  .onRequestTo(async request => {
-    const urlRegex = /^https:\/\/liveapi-cached.yext.com\/v2\/accounts\/me\/answers\/autocomplete/;
-    return urlRegex.test(request.url) && request.method === 'get';
-  })
-  .respond((req, res) => {
-    const parsedUrl = new URL(req.url);
-    res.body = JSON.stringify(generateAutoCompleteResponse(parsedUrl.searchParams.get('input')));
-    res.headers = CORSHeaders;
-    res.statusCode = 200;
-  });
